% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affect_donnees_externes.R
\name{affect_geol_plct_IFN}
\alias{affect_geol_plct_IFN}
\title{A partir d'un shapefile des placettes IFN auquelles des données géologiques ont été attribuées, cette fonction affecte
à chaque placette de la base de données IFN un type de regroupement géologique.}
\usage{
affect_geol_plct_IFN(IFN_data, shp_geol)
}
\arguments{
\item{IFN_data}{base de donnnées IFN (du même type que celle issue de la fonction import_IFN(...))}

\item{shp_geol}{shapefile des placettes IFN (avec positionnemeent en X et Y floutée) pour lequel sont spécifié:
l'identifiant de chaque placette (idp), un code de niveau de fertilite (niv_fertilite),
un classement entre acide/basique (basique_acide) et un classement entre roche dure/tendre (dure_tendre)}
}
\value{
renvoir la base de données IFN pour laquelle on a joint par placette et pour chaque arbre
les données géologiques : niveau de fertilité, acide/basique, dure/tendre
}
\description{
A partir d'un shapefile des placettes IFN auquelles des données géologiques ont été attribuées, cette fonction affecte
à chaque placette de la base de données IFN un type de regroupement géologique.
}
