% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_calibration_modele_accroissementG.R
\name{bd_calibration}
\alias{bd_calibration}
\title{Création d'une base de données par placette IFN contenant à la fois des variables dendrométriques et environementales}
\usage{
bd_calibration(IFN_data, code_ser = IFNtraitement::code_ser_num)
}
\arguments{
\item{IFN_data}{Base de données IFN issue du téléchargement par la fonction import_IFN(...) puis
traitée par l'ensemble des fonctions calc_dendro(...), extract_alt_plct_IFN(...), affect_geol_plct_IFN(...),
rayonnement(...) et precipitation(...),}

\item{code_ser}{tableau de correspondance entre le code de la sylvo-éco-région de l'IFN (codeser) et un code numérique qui permettra ensuite
l'application des modèles de prédictions à des données aux formats raster.}
}
\value{
renvoie un data_frame avec des données par placettes IFN comprenant des données "brute", issues des calculs dendrométriques,
de données ajoutées à partir de données cartographiées extérieur + ajout du code sylvo-écorégion, et d'un regroupement
en grandes classee de l'altitude, lapente et l'exposition.
}
\description{
Cette fonction renvoie une base de donnée par placette IFN comportant l'accroissement courant en surface terrière sur la placette
et un ensemble de données dendrométriques et environnementales.
Info : initialement ceci à été fait dans le but de calibrer un modèle d'accroissement en surface terrière.
}
