% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN_calculs_dendro.R
\name{calc_dendro}
\alias{calc_dendro}
\title{A partir des données brutes de l'IFN cette fonction calcul un ensemble de données dendrométriques à l'échelle de l'arbre et de la placette
(accroissement individuel d'un arbre, ... , surface terière totale de la placette, résineuse, ...)}
\usage{
calc_dendro(IFN_data, precomptable = 17.5,
  code_ess = IFNtraitement::code_ess)
}
\arguments{
\item{IFN_data}{données brutes de l'IFN téléchargées (cf. fonction import_IFN)}

\item{precomptable}{diamètre de précomptabilité pour effectuer les calculs dendrométriques. (Par défaut 17,5 cm)}

\item{code_ess}{tableau de correspondance entre les codes essences IFN (espar) et les codes essences ONF + regroupement par famille (R : résineux et F : feuillus )}
}
\value{
renvoie la base de données IFN avec l'ajout des calculs dendrométriques à l'échelle de l'arbre (cf. fonction calc_dendro_arbre)
et l'ajout sur la base de donnée placettes d'un ensemble de données dendrométriques à l'échelle placette :
densité de tige, surface terrière, accroissement en surface terrière, diamère quadratique moyen, surface terrière et accroissement
en surface terrière feuillus & résineuse, surface terrière de gros bois, porportion de feuillus, résineux et gros bois et
type de mélange entre feuillus et résieneux du peuplement)
}
\description{
A partir des données brutes de l'IFN cette fonction calcul un ensemble de données dendrométriques à l'échelle de l'arbre et de la placette
(accroissement individuel d'un arbre, ... , surface terière totale de la placette, résineuse, ...)
}
