% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN_calculs_dendro.R
\name{calc_dendro_arbre}
\alias{calc_dendro_arbre}
\title{Cette fonction calcule pour chaque arbre de l'inventaire foretier sa surface terrière individuelle
ainsi que son accroissement en diamètre et surface terrière}
\usage{
calc_dendro_arbre(IFN_data, precomptable = 17.5,
  code_ess = IFNtraitement::code_ess)
}
\arguments{
\item{IFN_data}{données brutes de l'IFN téléchargées (cf. fonction import_IFN)}

\item{precomptable}{diamètre de précomptabilité pour effectuer les calculs dendrométriques. (Par défaut 17,5 cm)}

\item{code_ess}{tableau de correspondance entre les codes essences IFN (espar) et les codes essences ONF + regroupement par famille (R : résineux et F : feuillus )}
}
\value{
renvoie la base de données IFN spécifiée en entrée avec l'ajout dans la base de données arbres des données :
surface terrière individuelle (g), accroissement individuel sur le diamètre (acc_d) et en surface terrière (acc_g)
}
\description{
Cette fonction calcule pour chaque arbre de l'inventaire foretier sa surface terrière individuelle
ainsi que son accroissement en diamètre et surface terrière
}
