% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN_calculs_dendro.R
\name{classif_cat_Diam}
\alias{classif_cat_Diam}
\title{Détermine pour chaque arbre de la base de données de l'IFN sa catégorie de diamètre en classe de 5cm à 5cm
et en grandes classes (perches, PB, BM, GB, TGB)}
\usage{
classif_cat_Diam(IFN_data)
}
\arguments{
\item{IFN_data}{données brutes de l'IFN téléchargées (cf. fonction import_IFN)}
}
\value{
renvoie la base de données IFN spécifiée en entrée avec deux champs supplémentaires dans la base de données arbres
(catégorie de diamètre de 5cm en 5cm et grande catégorie de diamètre)
}
\description{
Détermine pour chaque arbre de la base de données de l'IFN sa catégorie de diamètre en classe de 5cm à 5cm
et en grandes classes (perches, PB, BM, GB, TGB)
}
