% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affect_donnees_externes.R
\name{extract_alt_plct_IFN}
\alias{extract_alt_plct_IFN}
\title{Fonction qui à partir des coordonnées en X et Y floutées des données IFN crée un fichier de points de type shapefile
où chaque point correspond à une placette et leur affecte une altitude extraite d'un mnt spécifié en entée de la fonction}
\usage{
extract_alt_plct_IFN(IFN_data, mnt_L93, crs = 2154)
}
\arguments{
\item{IFN_data}{Base de données IFN au même format qu'à la sortie de la fonction import_IFN(...)}

\item{mnt_L93}{Modèle Numérique de Terrain (MNT / DEM), raster supposé être projeté en lambert 93}

\item{crs}{système de coordonnées géographiques de référence, par défaut : Lambert 93 code EPSG = 2154
actuellement seul ce system de coordonée fonctionne pour cette fonction}
}
\value{
retourne la base de données initial avec le rajout d'un élement à la liste de donnée correspondant
au spatialDataFrame des placettes auxquelles une altitude approximative à été affectée
}
\description{
Fonction qui à partir des coordonnées en X et Y floutées des données IFN crée un fichier de points de type shapefile
où chaque point correspond à une placette et leur affecte une altitude extraite d'un mnt spécifié en entée de la fonction
}
