% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fusion_BD_alti}
\alias{fusion_BD_alti}
\title{fonctions permettant la fusion de raster}
\usage{
fusion_BD_alti(list_dir_raster, nom_sortie, chemin_sortie = getwd(),
  ext = ".tif", crs = 2154)
}
\arguments{
\item{list_dir_raster}{liste comportant le chemin sur le disque dur de chaque raster à fussionner}

\item{nom_sortie}{nom du raster qui sera enregistrer en sortie}

\item{chemin_sortie}{chemin vers le dossier où sera enregistrer le raster en sortie de la fonction}

\item{ext}{extension selon laquelle le raster en sortie sera enregistré (par défaut ".tif")}

\item{crs}{système de coordonnées géographique de référence, par défaut : Lambert 93 (code EPSG = 2154)
Actuellement la fonction ne peut fonctionner qu'avec des raster en Lambert 93}
}
\value{
renvoie le raster une fois fussionné et l'enregistre dans les dossier et avec le nom spécifié en entrée de la fonction
}
\description{
initialement cette fonction à permit la fusion de BD_alti de l'IGN sur plusieurs département
}
