% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_IFN_data.R
\name{import_IFN}
\alias{import_IFN}
\title{Téléchargement en ligne de la base de données brutes de l'IFN}
\usage{
import_IFN(choix_dept = NULL, choix_ser = NULL, annees = NULL,
  split = T, interactif = F, save_dsn = "")
}
\arguments{
\item{choix_dept}{numéro du département ou liste des numéros de départements sur lequel /lesquels on souhaite obtenir les données
brutes de l'IFN. Par défaut vaut NULL et donc on récupère les données brutes de tous les départements métropolitain}

\item{choix_ser}{code de la sylvo-écorégion (format character/texte) ou liste des codes de la sylvo-écorégion sur laquel/lesquelles
on souhaite obtenir les données brutes de l'IFN. Par défaut vaut NULL et donc on récupère les données brutes de tous les ser}

\item{annees}{année (format numeric) ou liste d'années pour laquel/lesquels on souhaite télécharger les données brutes IFN
les données étant téléchargeable que depuis 2005, il n'est pas possible d'indiquer une année antérieure à 2005}

\item{split}{prend la valeur TRUE ou FALSE. Si la valeur est TRUE alors les données brutes IFN seront divisé en autant de data.frame
qu'il y a de sylvo-écorégion sur la zone où les données ont été téléchargé. Par défaut split vaut TRUE.}

\item{interactif}{vaut TRUE ou FALSE (FALSE par défaut), si vaut TRUE la fonction interragie avec l'utilisateur pour spécifier les paramètres
choix_dept, choix_ser et annees}

\item{save_dsn}{chemin complet pour l'enregistrement au format RDS des données IFN en sortie de la fonction, par défaut vaut "" et rien n'est enregistrer.
Si le chemein spécifié n'est pas correct le fichier n'est pas enregistrer (ex save_dsn = "C:/dos1/ssdos2/donnees_IFN" enregistrera
l'archive nommée donnees_IFN dans le sous dossier 2 du dossier 1 dans le répertoire C)}
}
\value{
la fonction retourne une liste avec trois sous listes correspondant chacune aux données placettes, arbres et écologies des données brutes de l'IFN
selon la valeur de split en entrée de la fonction chaque sous liste comportera 1 'si (split == FALSE) ou n dataframe correspondant chacun aux données restreintes
aux sylvo-éco-régions de la zone d'étude
}
\description{
Cette fonction télécharge en ligne les dernières données misent à jour de l'inventaire forestier national en pouvant remonter
jusqu'aux inventaires de l'année 2015 et en général va jusqu'à l'année actuelle - 2 ans. Il est possible de préciser une année ou
plage d'année sur laquelle on souhaite télécharger les données. Il est aussi possible de préciser les départements
(numéro des départements) et/ou des sylvo-éco-région (code des SER) pour lesquels on veut restraindre les données. les données pauvent même
être éclatées par sylvo-écorégion
}
