% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RMSE}
\alias{RMSE}
\title{fonction permettant le calcul de l'erreur quadratique moyenne entre données prédites et données de référence}
\usage{
RMSE(ref, predict)
}
\arguments{
\item{ref}{valeure ou vecteur des valeures de référence}

\item{predict}{valeure ou vecteur des valeures prédites}
}
\value{
renvoie l'ecart quadratique moyen entre valeures prédites et valeures de référence
}
\description{
les paramètres ref et predict doivent être de même longueur et bien ordonnée
(i.e. la nième valeur prédite doit correspondre à la nième valeur de référence)
}
