% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_IFN_data.R
\name{select_don_IFN}
\alias{select_don_IFN}
\title{Fonction interne à la fonction IFNtraitement::import_IFN(...) qui effectue le filtrage des données
selon les départements et sylvo-éco-régions spécifiées}
\usage{
select_don_IFN(IFN_data, choix_dept = NULL, choix_ser = NULL,
  split = T)
}
\arguments{
\item{choix_dept}{numéro du département ou liste des numéros de départements sur lequel /lesquels on souhaite obtenir les données
brutes de l'IFN. Par défaut vaut NULL et donc on récupère les données brutes de tous les départements métropolitain}

\item{choix_ser}{code de la sylvo-écorégion (format character/texte) ou liste des codes de la sylvo-écorégion sur laquel/lesquelles
on souhaite obtenir les données brutes de l'IFN. Par défaut vaut NULL et donc on récupère les données brutes de tous les ser}

\item{split}{prend la valeur TRUE ou FALSE. Si la valeur est TRUE alors les données brutes IFN seront divisé en autant de data.frame
qu'il y a de sylvo-écorégion sur la zone où les données ont été téléchargé. Par défaut split vaut TRUE.}

\item{annees}{année (format numeric) ou liste d'années pour laquel/lesquels on souhaite télécharger les données brutes IFN
les données étant téléchargeable que depuis 2005, il n'est pas possible d'indiquer une année antérieure à 2005}
}
\value{
la fonction retourne une liste avec trois sous listes correspondant chacune aux données placettes, arbres et écologies des données brutes de l'IFN
selon la valeur de split en entrée de la fonction chaque sous liste comportera 1 'si (split == FALSE) ou n dataframe correspondant chacun aux données restreindre aux sylvo-écorégion de la zone d'étude
}
\description{
Fonction interne à la fonction IFNtraitement::import_IFN(...) qui effectue le filtrage des données
selon les départements et sylvo-éco-régions spécifiées
}
