% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN_calculs_dendro.R
\name{select_plct_completes}
\alias{select_plct_completes}
\title{A partir des données brutes de l'IFN, cette fonction sélectionne les placettes pour lesquelless on possède
l'accroissement individuel de l'ensemble des arbres de la placettes}
\usage{
select_plct_completes(IFN_data)
}
\arguments{
\item{IFN_data}{données brutes de l'IFN téléchargées (cf. fonction import_IFN)}
}
\value{
renvoie la base de données filtrée sous le même format que la base de données spécifiée en entrée
}
\description{
A partir des données brutes de l'IFN, cette fonction sélectionne les placettes pour lesquelless on possède
l'accroissement individuel de l'ensemble des arbres de la placettes
}
